dataset-generation/nuget_data_generator/requirements.txt
aiohttp
jmespath
requests
typeguard
httpx
python-dateutil
aiofiles